Properly propagate exceptions through promise chain
Rethrowing an exception in the catch clause does not cause promise to
reach the "rejected" state. A warning would be logged in the output:

(node:11548) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.